requirements.txt
Django==4.2.2
uwsgi==2.0.21
requests==2.31.0
mysql==0.0.3
mysqlclient==2.1.1
MySQL-python==1.2.3
djangorestframework==3.14.0
Pygments==2.15.1
pytz==2023.3